+++ src/Articulate/Controllers/ArticulateRichTextController.cs
@@ -1,6 +1,8 @@
 #nullable enable
+using Articulate.Options;
 using Microsoft.AspNetCore.Mvc.ViewEngines;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Umbraco.Cms.Core.Models.PublishedContent;
 using Umbraco.Cms.Core.Web;
 
@@ -13,6 +15,7 @@ namespace Articulate.Controllers
         ILogger<ArticulateRichTextController> logger,
         ICompositeViewEngine compositeViewEngine,
         IUmbracoContextAccessor umbracoContextAccessor,
-        IPublishedValueFallback publishedValueFallback)
-        : BlogPostControllerBase(logger, compositeViewEngine, umbracoContextAccessor, publishedValueFallback);
+        IPublishedValueFallback publishedValueFallback,
+        IOptions<ArticulateCommentsOptions> commentsOptions)
+        : BlogPostControllerBase(logger, compositeViewEngine, umbracoContextAccessor, publishedValueFallback, commentsOptions);
 }