+++ src/Articulate/Models/ListModel.cs
@@ -1,4 +1,5 @@
 #nullable enable
+using Articulate.Options;
 using Umbraco.Cms.Core.Models.PublishedContent;
 
 namespace Articulate.Models
@@ -18,16 +19,18 @@ namespace Articulate.Models
         /// <param name="listItems"></param>
         /// <param name="pager"></param>
         /// <param name="publishedValueFallback"></param>
+        /// <param name="commentsOptions"></param>
         /// <remarks>
         /// Default sorting by published date will be disabled for this list model, it is assumed that the list items will
         /// already be sorted.
         /// </remarks>
         public ListModel(
-            IPublishedContent? content,
+            IPublishedContent content,
             PagerModel? pager,
             IEnumerable<IPublishedContent>? listItems,
-            IPublishedValueFallback publishedValueFallback)
-            : base(content, publishedValueFallback)
+            IPublishedValueFallback publishedValueFallback,
+            ArticulateCommentsOptions? commentsOptions = null)
+            : base(content, publishedValueFallback, commentsOptions)
         {
             ArgumentNullException.ThrowIfNull(content);
 
@@ -67,7 +70,7 @@ namespace Articulate.Models
             }
 
             return items
-                .Select(x => new PostModel(x, PublishedValueFallback))
+                .Select(x => new PostModel(x, PublishedValueFallback, CommentsOptions))
                 .ToArray();
         }
     }